refactor: Move "overlay" variant of slider to own component template
